requirements.txt
pandas
sqlalchemy
psycopg2-binary
streamlit
fastapi
uvicorn
httpx
python-dotenv